refactor WebMercatorViewport
