+++ modules/web-mercator/test/spec/web-mercator-viewport.spec.js
@@ -31,6 +31,21 @@ test('WebMercatorViewport#constructor - 0 width/height', t => {
   t.end();
 });
 
+test('WebMercatorViewport#equals', t => {
+  // TODO - fix types
+  // @ts-ignore
+  const viewport1 = new WebMercatorViewport(VIEWPORT_PROPS.flat);
+  // @ts-ignore
+  const viewport2 = new WebMercatorViewport(VIEWPORT_PROPS.flat);
+  // @ts-ignore
+  const viewport3 = new WebMercatorViewport(Object.assign({}, VIEWPORT_PROPS.flat, {height: 33}));
+
+  t.ok(viewport1.equals(viewport1), 'Viewport equality correct');
+  t.ok(viewport1.equals(viewport2), 'Viewport equality correct');
+  t.notOk(viewport1.equals(viewport3), 'Viewport equality correct');
+  t.end();
+});
+
 test('WebMercatorViewport.projectFlat', t => {
   config.EPSILON = 1e-6;
